Removed the old typechecker now that it is not needed.
