--- compiler/cx-compiler-typechecker-new/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "cx-compiler-typechecker-new"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-cx-data-lexer = { path = "../cx-data-lexer" }
-cx-data-typechecker = { path = "../cx-data-typechecker" }
-cx-data-ast = { path = "../cx-data-ast" }
-cx-data-pipeline = { path = "../cx-data-pipeline" }
-
-cx-util = { path = "../cx-util" }
