+++ compiler/cx-compiler-typechecker/Cargo.toml
@@ -1,16 +1,12 @@
 [package]
-name = "cx-compiler-typechecker"
-version = "0.0.0"
-edition = "2021"
-autotests = false
+name = "cx-compiler-typechecker-new"
+version = "0.1.0"
+edition = "2024"
 
 [dependencies]
-cx-data-ast = { path = "../cx-data-ast" }
-cx-data-bytecode = { path = "../cx-data-bytecode" }
 cx-data-lexer = { path = "../cx-data-lexer" }
 cx-data-typechecker = { path = "../cx-data-typechecker" }
-
-cx-compiler-ast = { path = "../cx-compiler-ast" }
+cx-data-ast = { path = "../cx-data-ast" }
+cx-data-pipeline = { path = "../cx-data-pipeline" }
 
 cx-util = { path = "../cx-util" }
-uuid = { version = "1.16.0", features = ["v4"] }
